Used this command:

    svn merge -r 1529475:HEAD https://svn.apache.org/repos/asf/roller/trunk .



git-svn-id: https://svn.apache.org/repos/asf/roller/branches/rome2@1546817 13f79535-47bb-0310-9956-ffa450edef68

+++ app/src/main/java/org/apache/roller/weblogger/pojos/BookmarkComparator.java
@@ -29,8 +29,8 @@ public class BookmarkComparator implements Comparator, Serializable
     {
         WeblogBookmark bd1 = (WeblogBookmark)val1;
         WeblogBookmark bd2 = (WeblogBookmark)val2;
-        int priority1 = bd1.getPriority().intValue();
-        int priority2 = bd2.getPriority().intValue();
+        int priority1 = bd1.getPriority();
+        int priority2 = bd2.getPriority();
 
         if (priority1 > priority2)
         {
